Refactoring and Research Unit tests
