+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/researching/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package holds everything connected to the {@link io.github.thebusybiscuit.slimefun4.core.researching.Research}
+ * class.
+ */
+package io.github.thebusybiscuit.slimefun4.core.researching;
